refactored api to use fastify plugin syntax
